% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QInference.R
\name{QFitInfer}
\alias{QFitInfer}
\title{Estimation and inference for individualized treatment rule using pooled-and-split de-correlated score}
\usage{
QFitInfer(qLearnFit, parallel = TRUE, indexToTest = c(1:8), intercept = TRUE)
}
\arguments{
\item{qLearnFit}{returns of \code{QLearnFit}}

\item{parallel}{whether use parallel computing; by default, FALSE.}

\item{indexToTest}{indicates which coefficients to test. By default, c(1:8)}

\item{intercept}{includes intercept or not}
}
\value{
p-values are the p-value for each coefficients included in indexToTest. (betaAN-1.96*sigmaAN/sqrt(sample size), betaAN+1.96*sigmaAN/sqrt(sample size)) provides the 95\% CI for the coefficients.
}
\description{
This function implements the Q-learning estimation for individualized treatment rule and the inference procedure based on the de-correlated score (see reference).
}
\references{
Muxuan Liang, Young-Geun Choi, Yang Ning, Maureen Smith, Yingqi Zhao (2020). Estimation and inference on high-dimensional individualized treatment rule in observational data using split-and-pooled de-correlated score.
}
\author{
Muxuan Liang <mliang@fredhutch.org>
}
