% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValueInfer.R
\name{ValueInfer}
\alias{ValueInfer}
\title{This function implements the inference for value of individualized treatment rule.}
\usage{
ValueInfer(
  data,
  method = "ITRFit",
  trainingfrac = 0.5 * log(NROW(data$predictor)),
  resamplingIter = 2000,
  ...
)
}
\arguments{
\item{data}{A list - list(predictor = x, treatment = trt, outcome = y)), where x is the covariate matrix, trt is 0 or 1 (1 indicates treatment), y is the outcome.}

\item{method}{method can be either 'ITRFit'  or 'QLearn'.}

\item{trainingfrac}{determines the traning sample size by \code{floor}(total sample size * \code{trainingfrac}).}

\item{resamplingIter}{determines how many resamplings. By default, it is 1000.}

\item{...}{Other parameters in \code{ITRFitInfer} and \code{QLearnFit}}
}
\value{
A list
\describe{
\item{value}{Estimation of the value function given the decision rule estimated by PEARL or Q-learning}
\item{se}{Estimated standard error of the value function}
\item{upper}{\code{value+1.96*se}}
\item{lower}{\code{value-1.96*se}}
}
}
\description{
This function implements the inference for value of individualized treatment rule.
}
\references{
Muxuan Liang, Young-Geun Choi, Yang Ning, Maureen Smith, Yingqi Zhao (2020). Estimation and inference on high-dimensional individualized treatment rule in observational data using split-and-pooled de-correlated score.
}
\author{
Muxuan Liang <mliang@fredhutch.org>
}
